refactor: 인증 관련 config, ui 클래스 패키지 이동
