+++ src/main/java/nextstep/member/application/TokenService.java
@@ -1,14 +1,14 @@
 package nextstep.member.application;
 
-import nextstep.member.AuthenticationException;
+import nextstep.auth.AuthenticationException;
 import nextstep.member.application.dto.TokenResponse;
 import nextstep.member.domain.Member;
 import org.springframework.stereotype.Service;
 
 @Service
 public class TokenService {
-    private MemberService memberService;
-    private JwtTokenProvider jwtTokenProvider;
+    private final MemberService memberService;
+    private final JwtTokenProvider jwtTokenProvider;
 
     public TokenService(MemberService memberService, JwtTokenProvider jwtTokenProvider) {
         this.memberService = memberService;
